Move sampling logic to a separate module
