+++ ml-dsa/src/algebra.rs
@@ -4,8 +4,7 @@ use hybrid_array::{
     Array,
 };
 
-use crate::crypto::{G, H};
-use crate::param::{ArraySize, Eta, MaskSamplingSize};
+use crate::param::ArraySize;
 use crate::util::Truncate;
 
 #[cfg(feature = "zeroize")]
@@ -29,48 +28,6 @@ impl FieldElement {
     pub const Q64: u64 = Self::Q as u64;
     pub const ONE: Self = Self(1);
     pub const MINUS_ONE: Self = Self(Self::Q - 1);
-
-    // Algorithm 14 CoeffFromThreeBytes
-    fn from_three_bytes(b: &[u8; 3]) -> Option<Self> {
-        let b0: Integer = b[0].into();
-        let b1: Integer = b[1].into();
-        let b2: Integer = b[2].into();
-
-        let b2p = if b2 > 127 { b2 - 128 } else { b2 };
-
-        let z = (b2p << 16) + (b1 << 8) + b0;
-        (z < Self::Q).then_some(FieldElement(z))
-    }
-
-    // Algorithm 15 CoeffFromHalfByte
-    fn from_half_byte(b: u8, eta: Eta) -> Option<Self> {
-        match eta {
-            Eta::Two if b < 15 => {
-                let b = (b as Integer) % 5;
-                if b <= 2 {
-                    Some(Self(2 - b))
-                } else {
-                    Some(-Self(b - 2))
-                }
-            }
-            Eta::Four if b < 9 => {
-                let b = b as Integer;
-                if b <= 4 {
-                    Some(Self(4 - b))
-                } else {
-                    Some(-Self(b - 4))
-                }
-            }
-            _ => None,
-        }
-    }
-
-    fn from_byte(z: u8, eta: Eta) -> (Option<Self>, Option<Self>) {
-        (
-            Self::from_half_byte(z & 0x0F, eta),
-            Self::from_half_byte(z >> 4, eta),
-        )
-    }
 
     fn mod_plus_minus(&self, m: Self) -> Self {
         let raw_mod = Self(self.0 % m.0);
@@ -207,13 +164,6 @@ impl Neg for FieldElement {
     }
 }
 
-// Algorithm 13 BytesToBits
-fn bit_set(z: &[u8], i: usize) -> bool {
-    let bit_index = i % 8;
-    let byte_index = i >> 3;
-    z[byte_index] & (1 << bit_index) != 0
-}
-
 /// An element of the ring `R_q`, i.e., a polynomial over `Z_q` of degree 255
 #[derive(Clone, Copy, Default, Debug, PartialEq)]
 pub struct Polynomial(pub Array<FieldElement, U256>);
@@ -233,66 +183,6 @@ impl Polynomial {
 
     fn infinity_norm(&self) -> u32 {
         self.0.iter().map(|x| x.infinity_norm()).max().unwrap()
-    }
-
-    // Algorithm 29 SampleInBall
-    pub fn sample_in_ball(rho: &[u8], tau: usize) -> Self {
-        const ONE: FieldElement = FieldElement(1);
-        const MINUS_ONE: FieldElement = FieldElement(FieldElement::Q - 1);
-
-        let mut c = Self::default();
-        let mut ctx = H::default().absorb(rho);
-
-        let mut s = [0u8; 8];
-        ctx.squeeze(&mut s);
-
-        // h == bytes_to_bits(s)
-        let mut j = [0u8];
-        for i in (256 - tau)..256 {
-            ctx.squeeze(&mut j);
-            while (j[0] as usize) > i {
-                ctx.squeeze(&mut j);
-            }
-
-            let j = j[0] as usize;
-            c.0[i] = c.0[j];
-            c.0[j] = if bit_set(&s, i + tau - 256) {
-                MINUS_ONE
-            } else {
-                ONE
-            };
-        }
-
-        c
-    }
-
-    // Algorithm 31 RejBoundedPoly
-    fn rej_bounded_poly(rho: &[u8], eta: Eta, r: u16) -> Self {
-        let mut j = 0;
-        let mut ctx = H::default().absorb(rho).absorb(&r.to_le_bytes());
-
-        let mut a = Self::default();
-        let mut z = [0u8];
-        while j < 256 {
-            ctx.squeeze(&mut z);
-            let (z0, z1) = FieldElement::from_byte(z[0], eta);
-
-            if let Some(z) = z0 {
-                a.0[j] = z;
-                j += 1;
-            }
-
-            if j == 256 {
-                break;
-            }
-
-            if let Some(z) = z1 {
-                a.0[j] = z;
-                j += 1;
-            }
-        }
-
-        a
     }
 
     // Algorithm 35 Power2Round
@@ -359,35 +249,6 @@ pub struct PolynomialVector<K: ArraySize>(pub Array<Polynomial, K>);
 impl<K: ArraySize> PolynomialVector<K> {
     pub fn mod_plus_minus(&self, m: FieldElement) -> Self {
         Self(self.0.iter().map(|x| x.mod_plus_minus(m)).collect())
-    }
-
-    // Algorithm 33 ExpandS
-    //
-    // We only do half of the algorithm here, because it's inconvenient to return two vectors of
-    // different sizes.  So the caller has to call twice:
-    //
-    //    let s1 = PolynomialVector::<K>::expand_s(rho, 0);
-    //    let s2 = PolynomialVector::<L>::expand_s(rho, L::USIZE);
-    pub fn expand_s(rho: &[u8], eta: Eta, base: usize) -> Self {
-        Self(Array::from_fn(|r| {
-            let r = (r + base) as u16;
-            Polynomial::rej_bounded_poly(rho, eta, r)
-        }))
-    }
-
-    pub fn expand_mask<Gamma1>(rho: &[u8], mu: u16) -> Self
-    where
-        Gamma1: MaskSamplingSize,
-    {
-        Self(Array::from_fn(|r| {
-            let r16: u16 = r.truncate();
-            let v = H::default()
-                .absorb(rho)
-                .absorb(&(mu + r16).to_le_bytes())
-                .squeeze_new::<Gamma1::SampleSize>();
-
-            Gamma1::unpack(&v)
-        }))
     }
 
     pub fn high_bits<Gamma2: Unsigned>(&self) -> Self {
@@ -463,25 +324,7 @@ impl<K: ArraySize> Mul<&PolynomialVector<K>> for FieldElement {
 #[derive(Clone, Default, Debug, PartialEq)]
 pub struct NttPolynomial(pub Array<FieldElement, U256>);
 
-impl NttPolynomial {
-    // Algorithm 30 RejNTTPoly
-    fn rej_ntt_poly(rho: &[u8], r: u8, s: u8) -> Self {
-        let mut j = 0;
-        let mut ctx = G::default().absorb(rho).absorb(&[s]).absorb(&[r]);
-
-        let mut a = Self::default();
-        let mut s = [0u8; 3];
-        while j < 256 {
-            ctx.squeeze(&mut s);
-            if let Some(x) = FieldElement::from_three_bytes(&s) {
-                a.0[j] = x;
-                j += 1;
-            }
-        }
-
-        a
-    }
-}
+impl NttPolynomial {}
 
 #[cfg(feature = "zeroize")]
 impl Zeroize for NttPolynomial {
@@ -520,72 +363,6 @@ impl Sub<&NttPolynomial> for &NttPolynomial {
         )
     }
 }
-
-/*
-// Algorithm 6. SampleNTT (lines 4-13)
-struct FieldElementReader<'a> {
-    xof: &'a mut dyn XofReader,
-    data: [u8; 96],
-    start: usize,
-    next: Option<Integer>,
-}
-
-impl<'a> FieldElementReader<'a> {
-    fn new(xof: &'a mut impl XofReader) -> Self {
-        let mut out = Self {
-            xof,
-            data: [0u8; 96],
-            start: 0,
-            next: None,
-        };
-
-        // Fill the buffer
-        out.xof.read(&mut out.data);
-
-        out
-    }
-
-    fn next(&mut self) -> FieldElement {
-        if let Some(val) = self.next {
-            self.next = None;
-            return FieldElement(val);
-        }
-
-        loop {
-            if self.start == self.data.len() {
-                self.xof.read(&mut self.data);
-                self.start = 0;
-            }
-
-            let end = self.start + 3;
-            let b = &self.data[self.start..end];
-            self.start = end;
-
-            let d1 = Integer::from(b[0]) + ((Integer::from(b[1]) & 0xf) << 8);
-            let d2 = (Integer::from(b[1]) >> 4) + ((Integer::from(b[2]) as Integer) << 4);
-
-            if d1 < FieldElement::Q {
-                if d2 < FieldElement::Q {
-                    self.next = Some(d2);
-                }
-                return FieldElement(d1);
-            }
-
-            if d2 < FieldElement::Q {
-                return FieldElement(d2);
-            }
-        }
-    }
-}
-
-impl NttPolynomial {
-    // Algorithm 6 SampleNTT(B)
-    pub fn sample_uniform(B: &mut impl XofReader) -> Self {
-        let mut reader = FieldElementReader::new(B);
-        Self(Array::from_fn(|_| reader.next()))
-    }
-}
-*/
 
 impl From<Array<FieldElement, U256>> for NttPolynomial {
     fn from(f: Array<FieldElement, U256>) -> NttPolynomial {
@@ -682,21 +459,9 @@ impl<K: ArraySize, L: ArraySize> Mul<&NttVector<L>> for &NttMatrix<K, L> {
     }
 }
 
-impl<K: ArraySize, L: ArraySize> NttMatrix<K, L> {
-    // Algorithm 32 ExpandA
-    pub fn expand_a(rho: &[u8]) -> Self {
-        Self(Array::from_fn(|r| {
-            NttVector(Array::from_fn(|s| {
-                NttPolynomial::rej_ntt_poly(rho, r as u8, s as u8)
-            }))
-        }))
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-    use hybrid_array::typenum::U16;
 
     #[test]
     fn polynomial_ops() {
@@ -706,66 +471,5 @@ mod test {
         assert_eq!((&f + &g), sum);
         assert_eq!((&sum - &g), f);
         assert_eq!(FieldElement(3) * &f, sum);
-    }
-
-    fn max_abs_1(p: Polynomial) -> bool {
-        p.0.iter()
-            .all(|x| x.0 == 0 || x.0 == 1 || x.0 == FieldElement::Q - 1)
-    }
-
-    fn hamming_weight(p: Polynomial) -> usize {
-        p.0.iter().filter(|x| x.0 != 0).count()
-    }
-
-    // Verify that SampleInBall returns a polynomial with the following properties:
-    //   a. All coefficients are from {-1, 0, 1}
-    //   b. Hamming weight is exactly tau
-    //
-    // We test 256 samples for each value of
-    #[test]
-    fn sample_in_ball() {
-        for tau in 1..65 {
-            for seed in 0..255 {
-                let rho = ((tau as u16) << 8) + (seed as u16);
-                let p = Polynomial::sample_in_ball(&rho.to_be_bytes(), tau);
-                assert_eq!(hamming_weight(p), tau);
-                assert!(max_abs_1(p));
-            }
-        }
-    }
-
-    // Verify that RejNTTPoly produces samples that are in the proper range, and roughly uniform.
-    // For the "roughly unform" criterion,
-    #[test]
-    fn rej_ntt_poly() {
-        let sample: Array<Array<FieldElement, U256>, U16> = Array::from_fn(|i| {
-            let i = i as u8;
-            let rho = [i; 32];
-            NttPolynomial::rej_ntt_poly(&rho, i, i + 1).into()
-        });
-
-        let sample = sample.as_flattened();
-
-        let all_in_range = sample.iter().all(|x| x.0 < FieldElement::Q);
-        assert!(all_in_range);
-
-        // TODO measure uniformity
-    }
-
-    #[test]
-    fn sample_cbd() {
-        let rho = [0; 32];
-
-        // Eta = 2
-        let sample = Polynomial::rej_bounded_poly(&rho, Eta::Two, 0).0;
-        let all_in_range = sample.iter().map(|x| *x + FieldElement(2)).all(|x| x.0 < 5);
-        assert!(all_in_range);
-        // TODO measure uniformity
-
-        // Eta = 4
-        let sample = Polynomial::rej_bounded_poly(&rho, Eta::Four, 0).0;
-        let all_in_range = sample.iter().map(|x| *x + FieldElement(4)).all(|x| x.0 < 9);
-        assert!(all_in_range);
-        // TODO measure uniformity
     }
 }
