+++ ml-dsa/src/sampling.rs
@@ -1,0 +1,238 @@
+use hybrid_array::Array;
+
+use crate::algebra::*;
+use crate::crypto::{G, H};
+use crate::param::{ArraySize, Eta, MaskSamplingSize};
+use crate::util::Truncate;
+
+// Algorithm 13 BytesToBits
+fn bit_set(z: &[u8], i: usize) -> bool {
+    let bit_index = i % 8;
+    let byte_index = i >> 3;
+    z[byte_index] & (1 << bit_index) != 0
+}
+
+// Algorithm 14 CoeffFromThreeBytes
+fn coeff_from_three_bytes(b: &[u8; 3]) -> Option<FieldElement> {
+    let b0: Integer = b[0].into();
+    let b1: Integer = b[1].into();
+    let b2: Integer = b[2].into();
+
+    let b2p = if b2 > 127 { b2 - 128 } else { b2 };
+
+    let z = (b2p << 16) + (b1 << 8) + b0;
+    (z < FieldElement::Q).then_some(FieldElement(z))
+}
+
+// Algorithm 15 CoeffFromHalfByte
+fn coeff_from_half_byte(b: u8, eta: Eta) -> Option<FieldElement> {
+    match eta {
+        Eta::Two if b < 15 => {
+            let b = (b as Integer) % 5;
+            if b <= 2 {
+                Some(FieldElement(2 - b))
+            } else {
+                Some(-FieldElement(b - 2))
+            }
+        }
+        Eta::Four if b < 9 => {
+            let b = b as Integer;
+            if b <= 4 {
+                Some(FieldElement(4 - b))
+            } else {
+                Some(-FieldElement(b - 4))
+            }
+        }
+        _ => None,
+    }
+}
+
+fn coeffs_from_byte(z: u8, eta: Eta) -> (Option<FieldElement>, Option<FieldElement>) {
+    (
+        coeff_from_half_byte(z & 0x0F, eta),
+        coeff_from_half_byte(z >> 4, eta),
+    )
+}
+
+// Algorithm 29 SampleInBall
+pub fn sample_in_ball(rho: &[u8], tau: usize) -> Polynomial {
+    const ONE: FieldElement = FieldElement(1);
+    const MINUS_ONE: FieldElement = FieldElement(FieldElement::Q - 1);
+
+    let mut c = Polynomial::default();
+    let mut ctx = H::default().absorb(rho);
+
+    let mut s = [0u8; 8];
+    ctx.squeeze(&mut s);
+
+    // h = bytes_to_bits(s)
+    let mut j = [0u8];
+    for i in (256 - tau)..256 {
+        ctx.squeeze(&mut j);
+        while (j[0] as usize) > i {
+            ctx.squeeze(&mut j);
+        }
+
+        let j = j[0] as usize;
+        c.0[i] = c.0[j];
+        c.0[j] = if bit_set(&s, i + tau - 256) {
+            MINUS_ONE
+        } else {
+            ONE
+        };
+    }
+
+    c
+}
+
+// Algorithm 30 RejNTTPoly
+fn rej_ntt_poly(rho: &[u8], r: u8, s: u8) -> NttPolynomial {
+    let mut j = 0;
+    let mut ctx = G::default().absorb(rho).absorb(&[s]).absorb(&[r]);
+
+    let mut a = NttPolynomial::default();
+    let mut s = [0u8; 3];
+    while j < 256 {
+        ctx.squeeze(&mut s);
+        if let Some(x) = coeff_from_three_bytes(&s) {
+            a.0[j] = x;
+            j += 1;
+        }
+    }
+
+    a
+}
+
+// Algorithm 31 RejBoundedPoly
+fn rej_bounded_poly(rho: &[u8], eta: Eta, r: u16) -> Polynomial {
+    let mut j = 0;
+    let mut ctx = H::default().absorb(rho).absorb(&r.to_le_bytes());
+
+    let mut a = Polynomial::default();
+    let mut z = [0u8];
+    while j < 256 {
+        ctx.squeeze(&mut z);
+        let (z0, z1) = coeffs_from_byte(z[0], eta);
+
+        if let Some(z) = z0 {
+            a.0[j] = z;
+            j += 1;
+        }
+
+        if j == 256 {
+            break;
+        }
+
+        if let Some(z) = z1 {
+            a.0[j] = z;
+            j += 1;
+        }
+    }
+
+    a
+}
+
+// Algorithm 32 ExpandA
+pub fn expand_a<K: ArraySize, L: ArraySize>(rho: &[u8]) -> NttMatrix<K, L> {
+    NttMatrix(Array::from_fn(|r| {
+        NttVector(Array::from_fn(|s| rej_ntt_poly(rho, r as u8, s as u8)))
+    }))
+}
+
+// Algorithm 33 ExpandS
+//
+// We only do half of the algorithm here, because it's inconvenient to return two vectors of
+// different sizes.  So the caller has to call twice:
+//
+//    let s1 = PolynomialVector::<K>::expand_s(rho, 0);
+//    let s2 = PolynomialVector::<L>::expand_s(rho, L::USIZE);
+pub fn expand_s<K: ArraySize>(rho: &[u8], eta: Eta, base: usize) -> PolynomialVector<K> {
+    PolynomialVector(Array::from_fn(|r| {
+        let r = (r + base) as u16;
+        rej_bounded_poly(rho, eta, r)
+    }))
+}
+
+// Algorithm 34 ExpandMask
+pub fn expand_mask<K, Gamma1>(rho: &[u8], mu: u16) -> PolynomialVector<K>
+where
+    K: ArraySize,
+    Gamma1: MaskSamplingSize,
+{
+    PolynomialVector(Array::from_fn(|r| {
+        let r16: u16 = r.truncate();
+        let v = H::default()
+            .absorb(rho)
+            .absorb(&(mu + r16).to_le_bytes())
+            .squeeze_new::<Gamma1::SampleSize>();
+
+        Gamma1::unpack(&v)
+    }))
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use hybrid_array::typenum::{U16, U256};
+
+    fn max_abs_1(p: Polynomial) -> bool {
+        p.0.iter()
+            .all(|x| x.0 == 0 || x.0 == 1 || x.0 == FieldElement::Q - 1)
+    }
+
+    fn hamming_weight(p: Polynomial) -> usize {
+        p.0.iter().filter(|x| x.0 != 0).count()
+    }
+
+    // Verify that SampleInBall returns a polynomial with the following properties:
+    //   a. All coefficients are from {-1, 0, 1}
+    //   b. Hamming weight is exactly tau
+    //
+    // We test 256 samples for each value of
+    #[test]
+    fn test_sample_in_ball() {
+        for tau in 1..65 {
+            for seed in 0..255 {
+                let rho = ((tau as u16) << 8) + (seed as u16);
+                let p = sample_in_ball(&rho.to_be_bytes(), tau);
+                assert_eq!(hamming_weight(p), tau);
+                assert!(max_abs_1(p));
+            }
+        }
+    }
+
+    // Verify that RejNTTPoly produces samples that are in the proper range, and roughly uniform.
+    // For the "roughly unform" criterion,
+    #[test]
+    fn test_rej_ntt_poly() {
+        let sample: Array<Array<FieldElement, U256>, U16> = Array::from_fn(|i| {
+            let i = i as u8;
+            let rho = [i; 32];
+            rej_ntt_poly(&rho, i, i + 1).into()
+        });
+
+        let sample = sample.as_flattened();
+
+        let all_in_range = sample.iter().all(|x| x.0 < FieldElement::Q);
+        assert!(all_in_range);
+
+        // TODO measure uniformity
+    }
+
+    #[test]
+    fn test_sample_cbd() {
+        let rho = [0; 32];
+
+        // Eta = 2
+        let sample = rej_bounded_poly(&rho, Eta::Two, 0).0;
+        let all_in_range = sample.iter().map(|x| *x + FieldElement(2)).all(|x| x.0 < 5);
+        assert!(all_in_range);
+        // TODO measure uniformity
+
+        // Eta = 4
+        let sample = rej_bounded_poly(&rho, Eta::Four, 0).0;
+        let all_in_range = sample.iter().map(|x| *x + FieldElement(4)).all(|x| x.0 < 9);
+        assert!(all_in_range);
+        // TODO measure uniformity
+    }
+}
